requirements.txt
pyyaml
h5py
pandas
numpy
opencv-python
tensorflow
keras
scikit-learn
kaggle
